Avoid warnings during compilation

+++ src/agg/include/agg_color_rgba.h
@@ -351,13 +351,13 @@ namespace agg
         //--------------------------------------------------------------------
         static AGG_INLINE double to_double(value_type a)
         {
-            return double(a) / base_mask;
+            return double(a) / double(base_mask);
         }
 
         //--------------------------------------------------------------------
         static AGG_INLINE value_type from_double(double a)
         {
-            return value_type(uround(a * base_mask));
+            return value_type(uround(a * double(base_mask)));
         }
 
         //--------------------------------------------------------------------
@@ -552,7 +552,7 @@ namespace agg
         AGG_INLINE self_type gradient(const self_type& c, double k) const
         {
             self_type ret;
-            calc_type ik = uround(k * base_mask);
+            calc_type ik = uround(k * double(base_mask));
             ret.r = lerp(r, c.r, ik);
             ret.g = lerp(g, c.g, ik);
             ret.b = lerp(b, c.b, ik);
@@ -743,13 +743,13 @@ namespace agg
         //--------------------------------------------------------------------
         static AGG_INLINE double to_double(value_type a)
         {
-            return double(a) / base_mask;
+            return double(a) / double(base_mask);
         }
 
         //--------------------------------------------------------------------
         static AGG_INLINE value_type from_double(double a)
         {
-            return value_type(uround(a * base_mask));
+            return value_type(uround(a * double(base_mask)));
         }
 
         //--------------------------------------------------------------------
@@ -944,7 +944,7 @@ namespace agg
         AGG_INLINE self_type gradient(const self_type& c, double k) const
         {
             self_type ret;
-            calc_type ik = uround(k * base_mask);
+            calc_type ik = uround(k * double(base_mask));
             ret.r = lerp(r, c.r, ik);
             ret.g = lerp(g, c.g, ik);
             ret.b = lerp(b, c.b, ik);
@@ -1181,9 +1181,9 @@ namespace agg
         }
 
         //--------------------------------------------------------------------
-        static AGG_INLINE value_type mult_cover(value_type a, cover_type b) 
+        static AGG_INLINE value_type mult_cover(value_type a, cover_type b)
         {
-            return value_type(a * b / cover_mask);
+            return value_type(a * b / float(cover_mask));
         }
 
         //--------------------------------------------------------------------
